requirements.txt
fastapi
uvicorn
pydantic
celery
redis
requests
python-dotenv
aiofiles
jinja2
python-multipart
sqlalchemy
psycopg2-binary
moviepy
ffmpeg-python
openai
Pillow
numpy
httpx
supabase
